Rename json to jsonflattener.
