--- plugins/transformer/json/.gitignore
@@ -1,1 +1,0 @@
-json
