--- plugins/transformer/json/docs/_configuration.md
@@ -1,9 +1,0 @@
-```yaml copy
-kind: transformer
-spec:
-  name: "json"
-  path: "cloudquery/json"
-  version: VERSION_TRANSFORMER_JSON
-  spec:
-    tables: ["xkcd_comics"]
-```
