+++ plugins/transformer/jsonflattener/.gitignore
@@ -1,0 +1,1 @@
+jsonflattener
