+++ plugins/transformer/jsonflattener/Makefile
