+++ plugins/transformer/jsonflattener/client/recordupdater/column_builders_timestamp.go
@@ -6,7 +6,7 @@ import (
 	"github.com/apache/arrow/go/v17/arrow"
 	"github.com/apache/arrow/go/v17/arrow/array"
 	"github.com/apache/arrow/go/v17/arrow/memory"
-	"github.com/cloudquery/cloudquery/plugins/transformer/json/client/schemaupdater"
+	"github.com/cloudquery/cloudquery/plugins/transformer/jsonflattener/client/schemaupdater"
 	"github.com/cloudquery/plugin-sdk/v4/types"
 )
