+++ plugins/transformer/jsonflattener/client/spec/spec.go
