+++ plugins/transformer/jsonflattener/client/spec/spec_test.go
