+++ plugins/transformer/jsonflattener/docs/_configuration.md
@@ -1,0 +1,9 @@
+```yaml copy
+kind: transformer
+spec:
+  name: "jsonflattener"
+  path: "cloudquery/jsonflattener"
+  version: VERSION_TRANSFORMER_JSONFLATTENER
+  spec:
+    tables: ["xkcd_comics"]
+```
