+++ plugins/transformer/jsonflattener/resources/plugin/plugin.go
@@ -2,7 +2,7 @@ package plugin
 
 // Don't move this file to a different package, it's used by Go releaser to embed the version in the binary.
 var (
-	Name    = "json"
+	Name    = "jsonflattener"
 	Kind    = "transformer"
 	Team    = "cloudquery"
 	Version = "development"
